detection_app/requirements.txt
torch
torchvision
pillow
gradio